requirements.txt
requests
rich
python-nmap
